requirements.txt
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.5.0
pyyaml>=6.0.1
